use audio progress for timing it runs SO MANY times per millisecond oh my god

+++ src/game.h
@@ -11,7 +11,7 @@ typedef enum {
 
 typedef struct {
     double s;
-    double track_ms;
+    int track_ms;
     double window_cooldown;
 } Time;
